Move Events#remove and Menus#remove into Repository;
- Handle multiple types of storage; Array, Hash and Set, better.
- Add more tests and placeholder tests.
- Add more documentation.

+++ lib/vedeu/support/repository.rb
@@ -9,7 +9,7 @@ module Vedeu
 
     # Return the whole repository.
     #
-    # @return [Hash]
+    # @return [Array|Hash|Set]
     def all
       storage
     end
@@ -39,7 +39,11 @@ module Vedeu
     #
     # @return [Array]
     def registered
-      storage.keys
+      return [] if storage.empty?
+
+      return storage.keys if storage.is_a?(Hash)
+
+      storage
     end
 
     # Returns a boolean indicating whether the named entity is registered.
@@ -47,7 +51,9 @@ module Vedeu
     # @param name [String]
     # @return [Boolean]
     def registered?(name)
-      storage.key?(name)
+      return false if storage.empty?
+
+      storage.include?(name)
     end
 
     # Returns the storage with the named entity removed, or false if the entity
@@ -56,9 +62,12 @@ module Vedeu
     # @param name [String]
     # @return [Hash|FalseClass]
     def remove(name)
+      return false if storage.empty?
+
       if registered?(name)
+        # Vedeu.log("Removing entity (#{entity}): '#{name}'")
         storage.delete(name)
-        storage
+        storage unless storage.is_a?(Set)
 
       else
         false
@@ -70,7 +79,7 @@ module Vedeu
 
     # Reset the repository.
     #
-    # @return [Hash]
+    # @return [Array|Hash|Set]
     def reset
       @_storage = in_memory
     end
